review comments

+++ src/redux/moviesActions.js
@@ -3,6 +3,7 @@ import * as movieService from '../service/MovieService';
 export const PENDING = 'PENDING';
 export const ERROR = 'ERROR';
 export const FETCH_MOVIES_SUCCESS = 'FETCH_MOVIES_SUCCESS';
+export const FETCH_MOVIES_PAGINATION_SUCCESS = 'FETCH_MOVIES_PAGINATION_SUCCESS';
 export const ADD_MOVIE_SUCCESS = 'ADD_MOVIE_SUCCESS';
 export const EDIT_MOVIE_SUCCESS = 'EDIT_MOVIE_SUCCESS';
 export const DELETE_MOVIE_SUCCESS = 'DELETE_MOVIE_SUCCESS';
@@ -27,12 +28,34 @@ function fetchMoviesSuccess(data) {
     });
 }
 
-export function fetchMovies() {
-    return (dispatch) => {
+function fetchMoviesPaginationSuccess(data) {
+    return dispatch => dispatch({
+        type: FETCH_MOVIES_PAGINATION_SUCCESS,
+        payload: data,
+    });
+}
+
+export function fetchMovies(sortBy, filter, search) {
+    return (dispatch, getState) => {
+        if (getState().movies.pending) return Promise.reject(new Error('Pending action'));
         dispatch(pending());
-        movieService.read()
+        return movieService.read(sortBy, filter, search)
             .then((res) => {
-                dispatch(fetchMoviesSuccess(res.data.data));
+                dispatch(fetchMoviesSuccess(res.data));
+            })
+            .catch((error) => {
+                dispatch(handleError(error?.response?.data.messages || error));
+            });
+    };
+}
+
+export function fetchMoviesPagination(sortBy, filter, search, offset) {
+    return (dispatch, getState) => {
+        if (getState().movies.pending) return Promise.reject(new Error('Pending action'));
+        dispatch(pending());
+        return movieService.read(sortBy, filter, search, offset)
+            .then((res) => {
+                dispatch(fetchMoviesPaginationSuccess(res.data));
             })
             .catch((error) => {
                 dispatch(handleError(error?.response?.data.messages || error));
@@ -44,7 +67,9 @@ export function fetchMoviesIfNeed() {
     return (dispatch, getState) => {
         const { movies } = getState();
         if (movies || movies === []) {
-            dispatch(fetchMovies());
+            return dispatch(fetchMovies());
+        } else {
+            return Promise.resolve();
         }
     };
 }
@@ -55,10 +80,12 @@ function addMovieSuccess(data) {
         payload: data,
     });
 }
+
 export function addMovie(data) {
-    return (dispatch) => {
+    return (dispatch, getState) => {
+        if (getState().movies.pending) return Promise.reject(new Error('Pending action'));
         dispatch(pending());
-        movieService.create(data)
+        return movieService.create(data)
             .then((res) => {
                 dispatch(addMovieSuccess(res.data));
             })
@@ -76,9 +103,10 @@ function editMovieSuccess(data) {
 }
 
 export function editMovie(data) {
-    return (dispatch) => {
+    return (dispatch, getState) => {
+        if (getState().movies.pending) return Promise.reject(new Error('Pending action'));
         dispatch(pending());
-        movieService.update(data)
+        return movieService.update(data)
             .then((res) => {
                 dispatch(editMovieSuccess(res.data));
             })
@@ -96,9 +124,10 @@ export function deleteMovieSuccess(movieId) {
 }
 
 export function deleteMovie(id) {
-    return (dispatch) => {
+    return (dispatch, getState) => {
+        if (getState().movies.pending) return Promise.reject(new Error('Pending action'));
         dispatch(pending());
-        movieService.remove(id)
+        return movieService.remove(id)
             .then(() => {
                 dispatch(deleteMovieSuccess(id));
             })
